# docs/swagger.yaml
basePath: /
definitions:
...
  input.ConfirmNewProfilePic:
    description: Chave da foto que foi enviada com sucesso para confirmar o upload
    properties:
      objectKey:
        description: Chave da foto no storage
        example: users/123/profile-pic.jpg
        type: string
    required:
    - objectKey
    type: object
...
  input.FriendshipRequestAnswer:
    description: Resposta para aceitar ou rejeitar uma solicitação de amizade pendente
    properties:
      answer:
        description: Resposta à solicitação (accept ou reject)
        enum:
        - accept
        - reject
        example: accept
        type: string
    required:
    - answer
    type: object
...
  input.RequestProfilePicUpdate:
    description: Dados para solicitar URL pré-assinada para upload de foto de perfil
    properties:
      contentType:
        description: Tipo de conteúdo da foto (image/jpeg, image/png, image/webp)
        enum:
        - image/jpeg
        - image/png
        - image/webp
        example: image/jpeg
        type: string
    required:
    - contentType
    type: object
...
  input.UpdateUser:
    description: Dados opcionais para atualizar informações do usuário autenticado
    properties:
      birthdate:
        description: Data de nascimento no formato YYYY-MM-DD (opcional)
        example: "1990-01-15"
        type: string
      name:
        description: Nome completo do usuário (opcional)
        example: João Silva
        type: string
      removeProfilePic:
        description: Flag para remover a foto de perfil (opcional)
        example: false
        type: boolean
    type: object
...
  response.Friend:
    description: Informações sobre um amigo do usuário autenticado
    properties:
      ProfilePicUrl:
        description: URL da foto de perfil do amigo (opcional)
        example: https://example.com/users/15/profile.jpg
        type: string
      friendsSince:
        description: Data em que se tornaram amigos (opcional)
        example: "2024-06-20T14:00:00Z"
        type: string
      id:
        description: ID do amigo
        example: 15
        type: integer
      name:
        description: Nome completo do amigo
        example: Maria Santos
        type: string
      nickname:
        description: Apelido do amigo
        example: mariasantos
        type: string
      since:
        description: Data de criação da conta do amigo
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  response.FriendshipRequest:
    description: Informações sobre uma solicitação de amizade pendente
    properties:
      id:
        description: ID do usuário que enviou a solicitação
        example: 42
        type: integer
      nickname:
        description: Apelido do usuário que enviou a solicitação
        example: joaosilva
        type: string
      profilePicUrl:
        description: URL da foto de perfil do usuário (opcional)
        example: https://example.com/users/42/profile.jpg
        type: string
    type: object
...
  response.GetUserResponse:
    description: Informações de um usuário retornado na busca/listagem
    properties:
      id:
        description: ID do usuário
        example: 42
        type: integer
      name:
        description: Nome completo do usuário
        example: João Silva
        type: string
      nickname:
        description: Apelido único do usuário
        example: joaosilva
        type: string
      profilePicUrl:
        description: URL da foto de perfil do usuário (se houver)
        example: https://example.com/users/42/profile.jpg
        type: string
      since:
        description: Data de criação da conta
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
...
  response.ProfilePicPresignedURL:
    description: URL pré-assinada e chave para upload de foto de perfil do usuário
    properties:
      objectKey:
        description: Chave da foto no storage
        example: users/123/profile-pic.jpg
        type: string
      presignedUrl:
        description: URL pré-assinada para upload da foto de perfil
        example: https://s3.amazonaws.com/bucket/users/123/profile-pic.jpg?X-Amz-Signature=...
        type: string
    type: object
...
  src_core_users_controller.ErrorMessageResponse:
    description: Resposta retornada quando ocorre um erro que não requer detalhes
      adicionais
    properties:
      message:
        type: string
    type: object
  src_core_users_controller.ValidationErrorResponse:
    description: Resposta retornada quando há erros de validação nos dados enviados
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API do Revit integrada com Swagger via Fiber.
  title: Revit API
  version: "1.0"
paths:
...
  /api/user:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de usuários com filtros opcionais
      parameters:
      - description: Limite de resultados por página (opcional, máximo 20)
        example: 10
        in: query
        maximum: 20
        name: limit
        type: integer
      - description: Apelido para busca (opcional, mínimo 3 caracteres)
        example: joao
        in: query
        minLength: 3
        name: nickname
        type: string
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de usuários
          schema:
            items:
              $ref: '#/definitions/response.GetUserResponse'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar usuários
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Atualiza os dados do usuário autenticado
      parameters:
      - description: Dados do usuário para atualização
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/input.UpdateUser'
      produces:
      - application/json
      responses:
        "204":
          description: Usuário atualizado com sucesso
        "400":
          description: 'Erro na validação dos dados. Possíveis campos: name, birthdate,
            removeProfilePic'
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Atualizar dados do usuário
      tags:
      - Users
  /api/user/{id}:
    get:
      consumes:
      - application/json
      description: Retorna os dados de um usuário específico
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Dados do usuário
          schema:
            $ref: '#/definitions/response.GetUserResponse'
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "404":
          description: Usuário não encontrado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Obter usuário por ID
      tags:
      - Users
  /api/user/email-available:
    get:
      consumes:
      - application/json
      description: Verifica se um email está disponível para uso
      parameters:
      - description: Email a ser verificado
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Email disponível
        "400":
          description: Erro na validação do email
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "409":
          description: Email já está em uso
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Verificar disponibilidade de email
      tags:
      - Users
  /api/user/friendship:
    get:
      consumes:
      - application/json
      description: Retorna a lista de amigos do usuário autenticado
      parameters:
      - description: Limite de resultados por página (opcional, máximo 20)
        example: 10
        in: query
        maximum: 20
        name: limit
        type: integer
      - description: Apelido para busca (opcional, mínimo 3 caracteres)
        example: joao
        in: query
        minLength: 3
        name: nickname
        type: string
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de amigos
          schema:
            items:
              $ref: '#/definitions/response.Friend'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar amigos
      tags:
      - Users
  /api/user/friendship/{destinataryId}:
    post:
      consumes:
      - application/json
      description: Envia uma solicitação de amizade para outro usuário
      parameters:
      - description: ID do usuário destinatário
        in: path
        name: destinataryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Solicitação de amizade enviada
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "404":
          description: Usuário destinatário não encontrado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Solicitar amizade
      tags:
      - Users
  /api/user/friendship/{friendId}:
    delete:
      consumes:
      - application/json
      description: Remove a amizade entre o usuário autenticado e outro usuário
      parameters:
      - description: ID do amigo a ser removido
        in: path
        name: friendId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Amizade removida com sucesso
        "400":
          description: ID inválido
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "404":
          description: Amizade não encontrada
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Remover amizade
      tags:
      - Users
  /api/user/friendship/{requesterId}:
    patch:
      consumes:
      - application/json
      description: Aceita ou rejeita uma solicitação de amizade recebida
      parameters:
      - description: ID do usuário que enviou a solicitação
        in: path
        name: requesterId
        required: true
        type: integer
      - description: Resposta à solicitação (accept/reject)
        in: body
        name: answer
        required: true
        schema:
          $ref: '#/definitions/input.FriendshipRequestAnswer'
      produces:
      - application/json
      responses:
        "204":
          description: Resposta processada com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "404":
          description: Solicitação não encontrada
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Responder solicitação de amizade
      tags:
      - Users
  /api/user/friendship/requests:
    get:
      consumes:
      - application/json
      description: Retorna as solicitações de amizade pendentes do usuário autenticado
      parameters:
      - description: Limite de resultados por página (opcional, máximo 20)
        example: 10
        in: query
        maximum: 20
        name: limit
        type: integer
      - description: Número da página (opcional, deve ser maior que 0)
        example: 1
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de solicitações de amizade
          schema:
            items:
              $ref: '#/definitions/response.FriendshipRequest'
            type: array
        "400":
          description: Erro na validação dos parâmetros
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Listar solicitações de amizade
      tags:
      - Users
  /api/user/nickname-available:
    get:
      consumes:
      - application/json
      description: Verifica se um nickname está disponível para uso
      parameters:
      - description: Nickname a ser verificado
        in: query
        name: nickname
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Nickname disponível
        "400":
          description: Erro na validação do nickname
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
        "409":
          description: Nickname já está em uso
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Verificar disponibilidade de nickname
      tags:
      - Users
  /api/user/profile-pic:
    patch:
      consumes:
      - application/json
      description: Confirma o upload de uma nova foto de perfil após o upload na URL
        pré-assinada
      parameters:
      - description: Confirmação do upload
        in: body
        name: confirmation
        required: true
        schema:
          $ref: '#/definitions/input.ConfirmNewProfilePic'
      produces:
      - application/json
      responses:
        "204":
          description: Foto de perfil confirmada com sucesso
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Confirmar nova foto de perfil
      tags:
      - Users
  /api/user/profile-pic/:
    post:
      consumes:
      - application/json
      description: Solicita uma URL pré-assinada para upload de nova foto de perfil
      parameters:
      - description: Dados da requisição de atualização de foto
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/input.RequestProfilePicUpdate'
      produces:
      - application/json
      responses:
        "201":
          description: URL pré-assinada para upload
          schema:
            $ref: '#/definitions/response.ProfilePicPresignedURL'
        "400":
          description: Erro na validação dos dados
          schema:
            $ref: '#/definitions/src_core_users_controller.ValidationErrorResponse'
        "401":
          description: Token inválido ou expirado
          schema:
            $ref: '#/definitions/src_core_users_controller.ErrorMessageResponse'
      security:
      - BearerAuth: []
      summary: Solicitar atualização de foto de perfil
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'Digite "Bearer " seguido do token JWT. Exemplo: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"